requirements.txt
# Workflow Engine 依赖

# 配置文件解析
PyYAML>=5.1

# 模板引擎（用于命令模板渲染）
Jinja2>=2.11

# HTTP请求（用于通知功能）
requests>=2.25

# 开发依赖（可选）
# pytest>=6.0
# pytest-cov>=2.10
# black>=21.0
# flake8>=3.8
